+++ .github/CHANGELOG.md
@@ -197,6 +197,7 @@
 1. [A380X/PRIM] Automatic THS trim after engine start or touch and go - @flogross89 (floridude)
 1. [A380X/PERF] Fixed issue where 0 was not an accepted cost index - @Jonny23787 (Jonathan)
 1. [A380X/MFD] Update MFD font to include question mark - @heclak (Heclak)
+1. [A32NX/ECP] Implemented realistic simulation of the ECAM control panel - @tracernz (Mike)
 
 ## 0.12.0